mv EnhancedContext React context to separate file
