+++ vscode/webviews/chat/EnhancedContext.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export const EnhancedContextEnabled: React.Context<boolean> = React.createContext(true)
+
+export function useEnhancedContextEnabled(): boolean {
+    return React.useContext(EnhancedContextEnabled)
+}
